Module backtrace::backtrace

Reexports

use std::fmt;

use std::os::raw::c_void;

use self::libunwind::trace as trace_imp;

use self::libunwind::Frame as FrameImp;

Modules

libunwind

Structs

Frame
A trait representing one frame of a backtrace, yielded to the trace function of this crate.

Functions

trace
Inspects the current call-stack, passing all active frames into the closure provided to calculate a stack trace.