Module backtrace::backtrace::libunwind::uw

Unwind library interface used for backtraces

Note that dead code is allowed as here are just bindings iOS doesn't use all of them it but adding more platform-specific configs pollutes the code too much

Reexports

pub use self::_Unwind_Reason_Code::*;

use libc;

use std::os::raw::c_int;

use std::os::raw::c_void;

Enums

_Unwind_Context

_Unwind_Reason_Code

Functions

_Unwind_Backtrace⚠

_Unwind_FindEnclosingFunction⚠

_Unwind_GetIPInfo⚠

Type Definitions

_Unwind_Trace_Fn