backtrace

Enum BytesOrWideString

pub enum BytesOrWideString<'a> {
    Bytes(&'a [u8]),
    Wide(&'a [u16]),
}

A platform independent representation of a string. When working with std enabled it is recommended to the convenience methods for providing conversions to std types.

Variants

Bytes(&'a [u8])

A slice, typically provided on Unix platforms.

Wide(&'a [u16])

Wide strings typically from Windows.

Implementations

impl<'a> BytesOrWideString<'a>

pub fn to_str_lossy(&self) -> Cow<'a, str>

Lossy converts to a Cow<str>, will allocate if Bytes is not valid UTF-8 or if BytesOrWideString is Wide.

Required features

This function requires the std feature of the backtrace crate to be enabled, and the std feature is enabled by default.

pub fn into_path_buf(self) -> PathBuf

Provides a Path representation of BytesOrWideString.

Required features

This function requires the std feature of the backtrace crate to be enabled, and the std feature is enabled by default.