backtrace

Struct BacktraceFmt

pub struct BacktraceFmt<'a, 'b> { /* private fields */ }

A formatter for backtraces.

This type can be used to print a backtrace regardless of where the backtrace itself comes from. If you have a Backtrace type then its Debug implementation already uses this printing format.

Implementations

impl<'a, 'b> BacktraceFmt<'a, 'b>

pub fn new( fmt: &'a mut Formatter<'b>, format: PrintFmt, print_path: &'a mut (dyn FnMut(&mut Formatter<'_>, BytesOrWideString<'_>) -> Result + 'b), ) -> Self

Create a new BacktraceFmt which will write output to the provided fmt.

The format argument will control the style in which the backtrace is printed, and the print_path argument will be used to print the BytesOrWideString instances of filenames. This type itself doesn’t do any printing of filenames, but this callback is required to do so.

pub fn add_context(&mut self) -> Result

Prints a preamble for the backtrace about to be printed.

This is required on some platforms for backtraces to be fully symbolicated later, and otherwise this should just be the first method you call after creating a BacktraceFmt.

pub fn frame(&mut self) -> BacktraceFrameFmt<'_, 'a, 'b>

Adds a frame to the backtrace output.

This commit returns an RAII instance of a BacktraceFrameFmt which can be used to actually print a frame, and on destruction it will increment the frame counter.

pub fn finish(&mut self) -> Result

Completes the backtrace output.

This is currently a no-op but is added for future compatibility with backtrace formats.

pub fn message(&mut self, msg: &str) -> Result

Inserts a message in the backtrace output.

This allows information to be inserted between frames, and won’t increment the frame_index unlike the frame method.

pub fn formatter(&mut self) -> &mut Formatter<'b>

Return the inner formatter.

This is used for writing custom information between frames with write! and writeln!, and won’t increment the frame_index unlike the frame method.