Struct backtrace::BacktraceSymbol

pub struct BacktraceSymbol {
    // some fields omitted
}

Captured version of a symbol in a backtrace.

This type is returned as a list from BacktraceFrame::symbols and represents the metadata for a symbol in a backtrace.