Struct backtrace::Frame

pub struct Frame { /* fields omitted */ }

A trait representing one frame of a backtrace, yielded to the trace function of this crate.

The tracing function's closure will be yielded frames, and the frame is virtually dispatched as the underlying implementation is not always known until runtime.

Methods

impl Frame

pub fn ip(&self) -> *mut c_void

Returns the current instruction pointer of this frame.

This is normally the next instruction to execute in the frame, but not all implementations list this with 100% accuracy (but it's generally pretty close).

It is recommended to pass this value to backtrace::resolve to turn it into a symbol name.

pub fn symbol_address(&self) -> *mut c_void

Returns the starting symbol address of the frame of this function.

This will attempt to rewind the instruction pointer returned by ip to the start of the function, returning that value. In some cases, however, backends will just return ip from this function.

The returned value can sometimes be used if backtrace::resolve failed on the ip given above.