Struct backtrace::SymbolName

pub struct SymbolName<'a> { /* private fields */ }

A wrapper around a symbol name to provide ergonomic accessors to the demangled name, the raw bytes, the raw string, etc.

Implementations

impl<'a> SymbolName<'a>

pub fn new(bytes: &'a [u8]) -> SymbolName<'a>

Creates a new symbol name from the raw underlying bytes.

pub fn as_str(&self) -> Option<&'a str>

Returns the raw (mangled) symbol name as a str if the symbol is valid utf-8.

Use the Display implementation if you want the demangled version.

pub fn as_bytes(&self) -> &'a [u8]

Returns the raw symbol name as a list of bytes