backtrace::symbolize::gimli::elf

Function locate_debugaltlink

fn locate_debugaltlink(
    path: &Path,
    filename: &OsStr,
    build_id: &[u8],
) -> Option<PathBuf>

Locate a file specified in a .gnu_debugaltlink section.

path is the file containing the section. filename and build_id are the contents of the section.

Search order is based on gdb:

filename, which is either absolute or relative to path

the build ID path under BUILD_ID_PATH

gdb also allows the user to customize the debug search path, but we don’t.

gdb also supports debuginfod, but we don’t yet.