Type Alias backtrace::symbolize::gimli::elf::Elf

type Elf = FileHeader64<NativeEndian>;

Aliased Type

struct Elf {

    pub e_ident: Ident,
    pub e_type: U16Bytes<LittleEndian>,
    pub e_machine: U16Bytes<LittleEndian>,
    pub e_version: U32Bytes<LittleEndian>,
    pub e_entry: U64Bytes<LittleEndian>,
    pub e_phoff: U64Bytes<LittleEndian>,
    pub e_shoff: U64Bytes<LittleEndian>,
    pub e_flags: U32Bytes<LittleEndian>,
    pub e_ehsize: U16Bytes<LittleEndian>,
    pub e_phentsize: U16Bytes<LittleEndian>,
    pub e_phnum: U16Bytes<LittleEndian>,
    pub e_shentsize: U16Bytes<LittleEndian>,
    pub e_shnum: U16Bytes<LittleEndian>,
    pub e_shstrndx: U16Bytes<LittleEndian>,

}

Fields
e_ident: Ident

Magic number and other information.

e_type: U16Bytes<LittleEndian>

Object file type. One of the ET_* constants.

e_machine: U16Bytes<LittleEndian>

Architecture. One of the EM_* constants.

e_version: U32Bytes<LittleEndian>

Object file version. Must be EV_CURRENT.

e_entry: U64Bytes<LittleEndian>

Entry point virtual address.

e_phoff: U64Bytes<LittleEndian>

Program header table file offset.

e_shoff: U64Bytes<LittleEndian>

Section header table file offset.

e_flags: U32Bytes<LittleEndian>

Processor-specific flags.

A combination of the EF_* constants.

e_ehsize: U16Bytes<LittleEndian>

Size in bytes of this header.

e_phentsize: U16Bytes<LittleEndian>

Program header table entry size.

e_phnum: U16Bytes<LittleEndian>

Program header table entry count.

If the count is greater than or equal to PN_XNUM then this field is set to PN_XNUM and the count is stored in the sh_info field of section 0.

e_shentsize: U16Bytes<LittleEndian>

Section header table entry size.

e_shnum: U16Bytes<LittleEndian>

Section header table entry count.

If the count is greater than or equal to SHN_LORESERVE then this field is set to 0 and the count is stored in the sh_size field of section 0. first section header.

e_shstrndx: U16Bytes<LittleEndian>

Section header string table index.

If the index is greater than or equal to SHN_LORESERVE then this field is set to SHN_XINDEX and the index is stored in the sh_link field of section 0.