backtrace::symbolize::gimli

Struct Mapping

struct Mapping {
    cx: Context<'static>,
    _map: Mmap,
    stash: Stash,
}

Fields
cx: Context<'static>_map: Mmapstash: Stash
Implementations

impl Mapping

pub fn new(path: &Path) -> Option<Mapping>

fn new_debug( original_path: &Path, path: PathBuf, crc: Option<u32>, ) -> Option<Mapping>

Load debuginfo from an external debug file.

fn load_dwarf_package<'data>( path: &Path, stash: &'data Stash, ) -> Option<Object<'data>>

Try to locate a DWARF package file.

impl Mapping

fn mk<F>(data: Mmap, mk: F) -> Option<Mapping>
where F: for<'a> FnOnce(&'a [u8], &'a Stash) -> Option<Context<'a>>,

Creates a Mapping by ensuring that the data specified is used to create a Context and it can only borrow from that or the Stash of decompressed sections or auxiliary data.

fn mk_or_other<F>(data: Mmap, mk: F) -> Option<Mapping>
where F: for<'a> FnOnce(&'a [u8], &'a Stash) -> Option<Either<Mapping, Context<'a>>>,

Creates a Mapping from data, or if the closure decides to, returns a different mapping.