Module backtrace::backtrace

Modules

libunwind

Traits

Frame
A trait representing one frame of a backtrace, yielded to the trace function of this crate.

Functions

trace
Inspects the current call-stack, passing all active frames into the closure provided to calculate a stack trace.