Module backtrace::backtrace

Modules

libunwind
Backtrace support using libunwind/gcc_s/etc APIs.

Structs

Frame
A trait representing one frame of a backtrace, yielded to the trace function of this crate.

Functions

trace
Inspects the current call-stack, passing all active frames into the closure provided to calculate a stack trace.

trace_unsynchronized⚠
Same as trace, only unsafe as it's unsynchronized.