backtrace

Function clear_symbol_cache

pub fn clear_symbol_cache()

Attempt to reclaim that cached memory used to symbolicate addresses.

This method will attempt to release any global data structures that have otherwise been cached globally or in the thread which typically represent parsed DWARF information or similar.

Caveats

While this function is always available it doesn’t actually do anything on most implementations. Libraries like dbghelp or libbacktrace do not provide facilities to deallocate state and manage the allocated memory. For now the std feature of this crate is the only feature where this function has any effect.