Function backtrace::resolve_frame_unsynchronized

pub unsafe fn resolve_frame_unsynchronized<F>(frame: &Frame, cb: F) where
    F: FnMut(&Symbol), 

Same as resolve_frame, only unsafe as it's unsynchronized.

This function does not have synchronization guarentees but is available when the std feature of this crate isn't compiled in. See the resolve_frame function for more documentation and examples.

Panics

See information on resolve_frame for caveats on cb panicking.