Function backtrace::resolve_unsynchronized

pub unsafe fn resolve_unsynchronized<F>(addr: *mut c_void, cb: F) where
    F: FnMut(&Symbol), 

Same as resolve, only unsafe as it’s unsynchronized.

This function does not have synchronization guarantees but is available when the std feature of this crate isn’t compiled in. See the resolve function for more documentation and examples.

Panics

See information on resolve for caveats on cb panicking.