Struct backtrace::Backtrace

pub struct Backtrace {
    // some fields omitted
}

Representation of an owned and self-contained backtrace.

This structure can be used to capture a backtrace at various points in a program and later used to inspect what the backtrace was at that time.

Methods

impl Backtrace

fn new() -> Backtrace

Captures a backtrace at the callsite of this function, returning an owned representation.

This function is useful for representing a backtrace as an object in Rust. This returned value can be sent across threads and printed elsewhere, and thie purpose of this value is to be entirely self contained.

Examples

use backtrace::Backtrace;

let current_backtrace = Backtrace::new();

fn frames(&self) -> &[BacktraceFrame]

Returns the frames from when this backtrace was captured.

The first entry of this slice is likely the function Backtrace::new, and the last frame is likely something about how this thread or the main function started.