Module backtrace::symbolize

Reexports

use std::fmt;

use std::marker::PhantomData;

use std::os::raw::c_void;

use std::path::Path;

use std::str;

use rustc_demangle::try_demangle;

use rustc_demangle::Demangle;

use self::libbacktrace::resolve as resolve_imp;

use self::libbacktrace::Symbol as SymbolImp;

Modules

libbacktrace

Structs

OptionCppSymbol

Symbol
A trait representing the resolution of a symbol in a file.

SymbolName
A wrapper around a symbol name to provide ergonomic accessors to the demangled name, the raw bytes, the raw string, etc.

Functions

resolve
Resolve an address to a symbol, passing the symbol to the specified closure.