Module backtrace::symbolize

Modules

gimli 🔒

Support for symbolication using the gimli crate on crates.io

Structs

Symbol

A trait representing the resolution of a symbol in a file.

SymbolName

A wrapper around a symbol name to provide ergonomic accessors to the demangled name, the raw bytes, the raw string, etc.

Enums

ResolveWhat

Functions

adjust_ip 🔒

clear_symbol_cache

Attempt to reclaim that cached memory used to symbolicate addresses.

format_symbol_name 🔒

resolve

Resolve an address to a symbol, passing the symbol to the specified closure.

resolve_frame

Resolve a previously capture frame to a symbol, passing the symbol to the specified closure.

resolve_frame_unsynchronized⚠

Same as resolve_frame, only unsafe as it’s unsynchronized.

resolve_unsynchronized⚠

Same as resolve, only unsafe as it’s unsynchronized.